feature: 스테이션 UI 및 데이터 연동 작업

+++ MPAndroidChart/app/src/main/java/com/dagger/mpandroidchart/model/ReportData.kt
@@ -3,5 +3,5 @@ package com.dagger.mpandroidchart.model
 import com.google.gson.annotations.SerializedName
 
 data class ReportData(
-    @SerializedName("socReport") var socReport: Map<String, Double>,
-    @SerializedName("countReport") var countReport: Map<String, Double>)
+    @SerializedName("socReport") var socReport: Map<String, Int>,
+    @SerializedName("countReport") var countReport: Map<String, Int>)
